requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
gql[all]==3.5.0
aiohttp==3.9.5
